// apps/web-svelte/src/lib/api/generated/model/managedControlPreview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ManagedControlState } from './managedControlState';
import type { ManagerOptions } from './managerOptions';

export interface ManagedControlPreview {
  scopeFingerprint: string;
  state: ManagedControlState;
  options: ManagerOptions;
}